[refactor] kakaoAPI 모델 수정
- 헤더 설정 시 안전한? 방법으로 변경
- endpoint 설정 시 쿼리 파라미터들 지정 가능하게 변경
